requirements.txt
fastapi
uvicorn
requests
chromadb
google.generativeai
